% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Estimate Percentages and 95 Percent Confidence Intervals in dplyr Pipelines}
\usage{
freq_table(.data, ..., percent_ci = 95, ci_type = "logit", drop = FALSE)
}
\arguments{
\item{.data}{A data frame. If it is already grouped (i.e., class == "grouped_df")
  then freq_table will ungroup it to prevent unexpected results.

  For two-way tables, the count for each level of the variable in the
  first argument to freq_table will be the denominator for row percentages
  and their 95% confidence intervals. Said another way, the goal of the
  analysis is to compare percentages of some characteristic across two or
  more groups of interest, then the variable in the first argument to
  freq_table should contain the groups of interest, and the variable in the
  second argument to freq_table should contain the characteristic of
  interest.}

\item{...}{Categorical variables to be used in calculations. Currently,
  freq_table accepts one or two variables -- not more.

  If "..." includes a factor variable with a level (category) that is
  unobserved in the data, that level will still appear in the results with
  a count (n) equal to zero. Further, the confidence intervals will be
  equal to NaN.}

\item{percent_ci}{sets the level, as a percentage, for confidence intervals.
The default is `percent_ci = 95` for 95% confidence intervals. The
percentage value entered (e.g., 95) is converted to an alpha level as
1 - (percent_ci / 100). It is then converted to a two-sided probability
as (1 - alpha / 2), which is used to calculate a critical value from
Student's t distribution with n - 1 degrees of freedom.}

\item{ci_type}{Selects the method used to estimate 95 percent confidence intervals.
The default for one-way and two-way tables is logit transformed ("log"). For
one-way tables only, ci_type can optionally calculate Wald ("linear")
confidence intervals using the "wald" argument.}

\item{drop}{If false (default) unobserved factor levels will be included in
the returned frequency table with an n of 0. For example, if you have a
factor variable, gender, but no males in your data then frequency table
returned by freq_table(df, gender) would still contain a row for
males with the variable n = 0. If drop is set to TRUE, then the resulting
frequency table would not include a row for males at all.}
}
\value{
A tibble with class "freq_table_one_way" or "freq_table_two_way"
}
\description{
The freq_table function produces one-way and two-way frequency
  tables for categorical variables. In addition to frequencies, the
  freq_table function displays percentages, and the standard errors and 95%
  confidence intervals of the percentages. For two-way tables only,
  freq_table also displays row (subgroup) percentages, standard errors,
  and 95 percent confidence intervals.

  freq_table is intended to be used in a dplyr pipeline.

  All standard errors are calculated as some version of:
  sqrt(proportion * (1 - proportion) / (n - 1))

  For one-way tables, the default 95 percent confidence intervals displayed are
  logit transformed confidence intervals equivalent to those used by Stata.
  Additionally, freq_table will return Wald ("linear") confidence intervals
  if the argument to ci_type = "wald".

  For two-way tables, freq_table returns logit transformed confidence
  intervals equivalent to those used by Stata.
}
\examples{
library(dplyr)
library(freqtables)

data(mtcars)

# One-way frequency table with defaults

mtcars \%>\%
  freq_table(am)

#> # A tibble: 2 x 7
#>     var   cat     n n_total percent   lcl   ucl
#>   <chr> <dbl> <int>   <int>   <dbl> <dbl> <dbl>
#> 1    am     0    19      32   59.38 40.94 75.50
#> 2    am     1    13      32   40.62 24.50 59.06

# Two-way frequency table with defaults

mtcars \%>\%
  freq_table(am, cyl)

#> # A tibble: 6 x 10
#>   row_var row_cat col_var col_cat     n n_row n_total percent_row lcl_row ucl_row
#>     <chr>   <dbl>   <chr>   <dbl> <int> <int>   <int>       <dbl>   <dbl>   <dbl>
#> 1      am       0     cyl       4     3    19      32       15.79    4.78   41.20
#> 2      am       0     cyl       6     4    19      32       21.05    7.58   46.44
#> 3      am       0     cyl       8    12    19      32       63.16   38.76   82.28
#> 4      am       1     cyl       4     8    13      32       61.54   32.30   84.29
#> 5      am       1     cyl       6     3    13      32       23.08    6.91   54.82
#> 6      am       1     cyl       8     2    13      32       15.38    3.43   48.18
}
\references{
Agresti, A. (2012). Categorical Data Analysis (3rd ed.). Hoboken, NJ: Wiley.

  SAS documentation: https://support.sas.com/documentation/cdl/en/statug/63347/HTML/default/viewer.htm#statug_surveyfreq_a0000000221.htm

  Stata documentation: https://www.stata.com/manuals13/rproportion.pdf
}
