% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_test.R
\name{freq_test}
\alias{freq_test}
\alias{freq_test.freq_table_one_way}
\alias{freq_test.freq_table_two_way}
\title{Hypothesis Testing for Frequency Tables}
\usage{
freq_test(.data, ...)

\method{freq_test}{freq_table_one_way}(.data, ...)

\method{freq_test}{freq_table_two_way}(.data, ...)
}
\arguments{
\item{.data}{A tibble of class freq_table_one_way or freq_table_two_way.}

\item{...}{Other parameters to be passed on.}
}
\value{
A tibble.
}
\description{
The freq_test function is an S3 generic. It currently has
  methods for conducting hypothesis tests on one-way and two-way frequency
  tables. It is designed to work in a dplyr pipeline with the freq_table
  function.

  For the freq_table_one_way class, the method used is Pearson's chi-square
  test for equality of proportions.

  For the freq_table_two_way class, the methods used are Pearson's
  chi-square test of independence and Fisher's exact test (when any
  expected cell count is <= 5).
}
\examples{
library(dplyr)
library(freqtables)

data(mtcars)

# Test equality of proportions
mtcars \%>\%
  freq_table(am) \%>\%
  freq_test()

# Chi-square test of independence
mtcars \%>\%
  group_by(cyl) \%>\%
  freq_table(am) \%>\%
  freq_test()
}
